Complete rewrite of rendering to fix fish-eye
Now we are back to drawing the wall first and then filling the rest.
The difference is that I worked out how to get the distance travelled
for every y to be drawn, so the ceiling and floor are consistent with
the walls.

Unfortunately some features were lost (exponential marching), the underside of walls have some weird white bar, and parts of the floor are just a black line. Still better than before.